requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
deepl==1.19.1
discord.py==2.0.1
distro==1.9.0
docutils==0.20.1
easyocr==1.7.1
filelock==3.16.0
firebase-admin==6.5.0
Flask==3.0.3
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.19.2
google-api-python-client==2.52.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-language==2.14.0
google-cloud-storage==2.18.2
google-cloud-vision==3.7.4
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.8
imageio==2.35.1
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.3
ninja==1.11.1.1
numpy==2.1.1
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pdfkit==1.0.0
pdfservices-sdk==4.0.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
selenium==4.1.0
setuptools==69.5.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sympy==1.13.2
tifffile==2024.8.30
tokenizers==0.20.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
urllib3-secure-extra==0.1.0
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.11.1
zipfile36==0.1.3
